refactor: inference results split into submodule
